requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
python-dotenv==1.0.0
boto3==1.34.0
psycopg2-binary==2.9.9
Flask-CORS==4.0.0
flask-socketio==5.3.4
eventlet==0.33.3
amazon-transcribe==0.6.2
pydub==0.25.1
hypothesis==6.92.0
pytest==7.4.3
pytest-asyncio==0.21.1
gunicorn==21.2.0
pydantic==2.10.5
reportlab==4.0.7
APScheduler==3.10.4
